Support x86_64.

+++ src/ubcsat-io.c
@@ -30,8 +30,8 @@
 #define MERSENNE
 
 #ifdef MERSENNE
-  extern unsigned long genrand_int32();
-  extern void init_genrand(unsigned long s);
+  extern UINT32 genrand_int32();
+  extern void init_genrand(UINT32 s);
 
   FXNRAND32 fxnRandUInt32 = genrand_int32;
   #define fxnRandSeed(A) init_genrand(A)
